internal: Extract underlying createStubInstance
