+++ lib/sinon/sandbox.js
@@ -11,6 +11,7 @@ var sinonClock = require("./util/fake-timers");
 var sinonMock = require("./mock");
 var sinonSpy = require("./spy");
 var sinonStub = require("./stub");
+var sinonCreateStubInstance = require("./create-stub-instance");
 var sinonFake = require("./fake");
 var valueToString = require("@sinonjs/commons").valueToString;
 var fakeServer = require("nise").fakeServer;
@@ -71,7 +72,7 @@ function Sandbox() {
     };
 
     sandbox.createStubInstance = function createStubInstance() {
-        var stubbed = sinonStub.createStubInstance.apply(null, arguments);
+        var stubbed = sinonCreateStubInstance.apply(null, arguments);
 
         var ownMethods = collectOwnMethods(stubbed);
