+++ lib/sinon/stub.js
@@ -126,33 +126,6 @@ function stub(object, property) {
     return isStubbingNonFuncProperty ? s : wrapMethod(object, property, s);
 }
 
-stub.createStubInstance = function (constructor, overrides) {
-    if (typeof constructor !== "function") {
-        throw new TypeError("The constructor should be a function.");
-    }
-
-    const stubInstance = Object.create(constructor.prototype);
-    stubInstance[Symbol.for("SinonType")] = "StubInstance";
-
-    const stubbedObject = stub(stubInstance);
-
-    forEach(Object.keys(overrides || {}), function (propertyName) {
-        if (propertyName in stubbedObject) {
-            var value = overrides[propertyName];
-            if (value && value.createStubInstance) {
-                stubbedObject[propertyName] = value;
-            } else {
-                stubbedObject[propertyName].returns(value);
-            }
-        } else {
-            throw new Error(
-                `Cannot stub ${propertyName}. Property does not exist!`
-            );
-        }
-    });
-    return stubbedObject;
-};
-
 function assertValidPropertyDescriptor(descriptor, property) {
     if (!descriptor || !property) {
         return;
